requirements.txt
things.py
